+++ api/src/modules/email/dtos/create-authorized-email-in.dto.ts
@@ -1,9 +1,0 @@
-import { IsEmail, IsString } from 'class-validator';
-
-export class CreateAuthorizedEmailInDto {
-
-  @IsString()
-  @IsEmail()
-  email: string;
-
-}